# feat: implement core HRV data structures and validation (issue #82)