## Summary

Implement comprehensive Heart Rate Variability (HRV) data structures and validation logic to support recovery tracking and athlete monitoring, addressing all requirements from issue #82.

## Changes

This PR adds a new `recovery` module (`src/recovery.rs`) with core HRV infrastructure:

### 📊 Data Structures

**HrvStatus Enum** (4 variants)
- `Poor` - HRV >30% below baseline, inadequate recovery
- `Unbalanced` - HRV 15-30% below baseline, partial recovery  
- `Balanced` - HRV within ±15% of baseline, good recovery
- `NoReading` - Insufficient data or measurement error

**HrvMetrics Struct**
- `rmssd: Option<f64>` - Root mean square of successive differences (ms)
- `status: Option<HrvStatus>` - Recovery status category
- `baseline: Option<f64>` - Personal baseline RMSSD (ms)
- `score: Option<u8>` - Normalized score 0-100
- `measurement_time: Option<DateTime<Utc>>` - Timestamp
- `measurement_context: Option<String>` - Context (sleep, morning, pre-workout)

**HrvMeasurement Struct**
- Complete record for individual HRV readings
- Athlete ID, timestamp, RMSSD, baseline, status, score
- Measurement context and source tracking
- Extensible metadata field (JSON)

**HrvValidationError Enum**
- Type-safe error handling
- Descriptive error messages with suggestions
- Invalid RMSSD, baseline, and score variants

### 🔬 Sports Science Implementation

**Status Determination Algorithm**
```rust
fn from_rmssd(rmssd: f64, baseline: f64) -> HrvStatus {
    let deviation_pct = ((rmssd - baseline) / baseline) * 100.0;
    
    if deviation_pct >= -15.0 { Balanced }
    else if deviation_pct >= -30.0 { Unbalanced }
    else { Poor }
}
```

**RMSSD Validation**
- Valid physiological range: 10-200ms
- Elite athletes: 60-100ms+
- Average adults: 20-50ms
- Low fitness/high stress: <20ms

**Score Calculation**
- With baseline: Ratio-based (rmssd/baseline × 100)
- Without baseline: Absolute scale (20ms=20, 100ms=100)
- Clamped to 0-100 range

**Baseline Updates**
- Exponential moving average (EMA)
- Smoothing factor α (typically 0.1-0.3)
- Formula: `new_baseline = α × rmssd + (1-α) × old_baseline`

### 📚 Documentation

**Comprehensive Module Documentation**
- Sports science background on HRV and autonomic nervous system
- RMSSD interpretation guidelines with typical ranges
- Status categories explained with physiological context
- Measurement timing best practices
- Personal baseline calculation methods

**Code-Level Documentation**
- Detailed function/struct documentation
- Algorithm explanations with formulas
- Usage examples in doc comments
- Parameter descriptions and return values

### ✅ Validation & Error Handling

**Input Validation**
- RMSSD range: 10-200ms (catches measurement errors)
- Baseline validation with same constraints
- Score validation: 0-100 range
- Type-safe error propagation

**Error Messages**
```rust
InvalidRmssd(45.5) → "Invalid RMSSD value: 45.5ms (valid range: 10-200ms)"
InvalidBaseline(250.0) → "Invalid baseline value: 250.0ms (valid range: 10-200ms)"
```

### 🔄 Serialization

**Full Serde Support**
- JSON serialization/deserialization
- Binary format support via serde
- All structs derive Serialize/Deserialize
- Round-trip integrity tested

### 🧪 Testing

**11 Comprehensive Unit Tests** (100% coverage)

1. `test_hrv_status_from_rmssd` - Status calculation accuracy
2. `test_hrv_status_display` - Display formatting
3. `test_valid_rmssd_range` - Range validation boundaries
4. `test_hrv_metrics_creation` - Struct construction
5. `test_hrv_metrics_invalid_rmssd` - Error handling for invalid RMSSD
6. `test_hrv_metrics_invalid_baseline` - Error handling for invalid baseline
7. `test_hrv_score_calculation` - Score algorithm correctness
8. `test_baseline_update` - EMA baseline updates
9. `test_hrv_measurement_creation` - Measurement struct
10. `test_hrv_measurement_serialization` - JSON round-trip
11. `test_hrv_status_transitions` - All status boundary conditions

**Test Results**: ✅ 11 passed, 0 failed, 0 warnings

## Acceptance Criteria

All requirements from issue #82 completed:

### Data Structures
- [x] HrvStatus enum (Poor, Unbalanced, Balanced, NoReading)
- [x] HrvMetrics struct with RMSSD, baseline, score, status
- [x] HrvMeasurement struct for individual readings
- [x] Serialization/deserialization support
- [x] Validation for HRV value ranges

### Testing
- [x] Unit tests for HRV status transitions (11 tests)
- [x] Validation tests for RMSSD ranges (10-200ms)
- [x] Serialization round-trip tests

### Documentation
- [x] Sports science background in module docs
- [x] RMSSD interpretation guidelines
- [x] Status categories with physiological context
- [x] Algorithm documentation with formulas
- [x] Usage examples in doc comments

### Quality
- [x] All structures compile without errors
- [x] Serialization to/from JSON works correctly
- [x] Value validation prevents invalid measurements
- [x] No warnings in module code
- [x] Type-safe error handling

## Implementation Notes

**Module Organization**
- Created new `src/recovery.rs` module for HRV functionality
- Added to `src/lib.rs` for library exposure
- Follows existing module patterns (similar to `training_effect`, `pmc`)

**Future Extensions**
- Ready for database integration (HrvMeasurement has id/athlete_id fields)
- Extensible metadata field for device-specific data
- Context tracking enables measurement consistency analysis
- Baseline calculation can be enhanced with rolling window logic

## Related Issues

Part of parent issue #45 (Recovery & HRV Metrics from FIT Files)
Closes #82

## Testing Instructions

```bash
# Run HRV module tests
cargo test recovery::tests

# Build library
cargo build --lib

# Check for warnings
cargo clippy -- -D warnings
```

🤖 Generated with [Claude Code](https://claude.com/claude-code)